Linked to 1.5.68

+++ DemoPlugins/src/org/bimserver/demoplugins/service/planner/EventLog.java
@@ -184,9 +184,9 @@ public class EventLog implements Iterable<Event> {
 
 	private GregorianCalendar parseDate(IfcDateAndTime start) {
 		GregorianCalendar gregorianCalendar = new GregorianCalendar();
-		gregorianCalendar.set(Calendar.YEAR, start.getDateComponent().getYearComponent());
-		gregorianCalendar.set(Calendar.MONTH, start.getDateComponent().getMonthComponent() - 1);
-		gregorianCalendar.set(Calendar.DAY_OF_MONTH, start.getDateComponent().getDayComponent());
+		gregorianCalendar.set(Calendar.YEAR, (int) start.getDateComponent().getYearComponent());
+		gregorianCalendar.set(Calendar.MONTH, (int) (start.getDateComponent().getMonthComponent() - 1));
+		gregorianCalendar.set(Calendar.DAY_OF_MONTH, (int) start.getDateComponent().getDayComponent());
 		return gregorianCalendar;
 	}
